--- ivy/functional/backends/tensorflow/extensions.py
@@ -1,242 +1,0 @@
-from typing import Union, Optional, Tuple, Literal, Sequence
-from numbers import Number
-import ivy
-from ivy.functional.ivy.extensions.sparse_array import (
-    _verify_coo_components,
-    _verify_csr_components,
-    _is_data_not_indices_values_and_shape,
-    _is_coo_not_csr,
-)
-from ivy.func_wrapper import with_unsupported_dtypes
-from . import backend_version
-import tensorflow as tf
-import tensorflow_probability as tfp
-import logging
-
-
-def is_native_sparse_array(x):
-    return isinstance(x, tf.SparseTensor)
-
-
-def native_sparse_array(
-    data=None,
-    *,
-    coo_indices=None,
-    csr_crow_indices=None,
-    csr_col_indices=None,
-    values=None,
-    dense_shape=None,
-):
-    if _is_data_not_indices_values_and_shape(
-        data, coo_indices, csr_crow_indices, csr_col_indices, values, dense_shape
-    ):
-        ivy.assertions.check_true(
-            ivy.is_native_sparse_array(data), message="not a sparse array"
-        )
-        return data
-    elif _is_coo_not_csr(
-        coo_indices, csr_crow_indices, csr_col_indices, values, dense_shape
-    ):
-        _verify_coo_components(
-            indices=coo_indices, values=values, dense_shape=dense_shape
-        )
-        all_coordinates = []
-        for i in range(values.shape[0]):
-            coordinate = ivy.gather(coo_indices, ivy.array([[i]]))
-            coordinate = ivy.reshape(coordinate, (coo_indices.shape[0],))
-            all_coordinates.append(coordinate.to_list())
-        return tf.SparseTensor(
-            indices=all_coordinates, values=values, dense_shape=dense_shape
-        )
-    else:
-        _verify_csr_components(
-            crow_indices=csr_crow_indices,
-            col_indices=csr_col_indices,
-            values=values,
-            dense_shape=dense_shape,
-        )
-        logging.warning(
-            "Tensorflow does not support CSR sparse array natively. None is returned."
-        )
-        return None
-
-
-def native_sparse_array_to_indices_values_and_shape(x):
-    if isinstance(x, tf.SparseTensor):
-        return x.indices, x.values, x.dense_shape
-    raise ivy.exceptions.IvyException("not a SparseTensor")
-
-
-def vorbis_window(
-    window_length: Union[tf.Tensor, tf.Variable],
-    *,
-    dtype: Optional[tf.DType] = tf.dtypes.float32,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    return tf.signal.vorbis_window(window_length, dtype=dtype, name=None)
-
-
-@with_unsupported_dtypes(
-    {"2.9.1 and below": ("uint8", "uint16", "uint32", "uint64")}, backend_version
-)
-def lcm(
-    x1: Union[tf.Tensor, tf.Variable],
-    x2: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    if [x1.dtype, x2.dtype] == [tf.int8, tf.int8]:
-        dtype = tf.int8
-        x1 = tf.cast(x1, dtype=tf.int16)
-        x2 = tf.cast(x2, dtype=tf.int16)
-    else:
-        dtype = x1.dtype
-    return tf.math.abs(tf.cast(tf.experimental.numpy.lcm(x1, x2), dtype=dtype))
-
-
-def hann_window(
-    window_length: int,
-    periodic: Optional[bool] = True,
-    dtype: Optional[tf.DType] = None,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    return tf.signal.hann_window(
-        window_length, periodic=periodic, dtype=dtype, name=None
-    )
-
-
-def max_pool2d(
-    x: Union[tf.Tensor, tf.Variable],
-    kernel: Union[int, Tuple[int], Tuple[int, int]],
-    strides: Union[int, Tuple[int], Tuple[int, int]],
-    padding: str,
-    /,
-    *,
-    data_format: str = "NHWC",
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    if data_format == "NCHW":
-        x = tf.transpose(x, (0, 2, 3, 1))
-    res = tf.nn.max_pool2d(x, kernel, strides, padding)
-    if data_format == "NCHW":
-        return tf.transpose(res, (0, 3, 1, 2))
-    return res
-
-
-def pad(
-    x: tf.Tensor,
-    /,
-    pad_width: tf.Tensor,
-    *,
-    mode: Optional[Literal["constant", "reflect", "symmetric"]] = "constant",
-    stat_length: Optional[Union[tf.Tensor, int]] = None,
-    constant_values: Optional[Number] = 0,
-    end_values: Optional[Number] = 0,
-    reflect_type: Optional[Literal["even", "odd"]] = "even",
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> tf.Tensor:
-    if x.shape == ():
-        x = tf.reshape(x, (-1,))
-    if mode == "constant":
-        return tf.pad(
-            x,
-            pad_width,
-            mode=mode,
-            constant_values=constant_values,
-        )
-    else:
-        return tf.pad(x, pad_width, mode=mode)
-
-
-def kaiser_window(
-    window_length: int,
-    periodic: bool = True,
-    beta: float = 12.0,
-    *,
-    dtype: Optional[tf.DType] = None,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    if periodic is False:
-        return tf.signal.kaiser_window(
-            window_length, beta, dtype=tf.dtypes.float32, name=None
-        )
-    else:
-        return tf.signal.kaiser_window(window_length + 1, beta, dtype=dtype, name=None)[
-            :-1
-        ]
-
-
-def moveaxis(
-    a: Union[tf.Tensor, tf.Variable],
-    source: Union[int, Sequence[int]],
-    destination: Union[int, Sequence[int]],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    return tf.experimental.numpy.moveaxis(a, source, destination)
-
-
-@with_unsupported_dtypes({"2.9.1 and below": ("bfloat16",)}, backend_version)
-def heaviside(
-    x1: Union[tf.Tensor, tf.Variable],
-    x2: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    return tf.experimental.numpy.heaviside(x1, x2)
-
-
-def median(
-    input: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    axis: Optional[Union[Tuple[int], int]] = None,
-    keepdims: Optional[bool] = False,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    return tfp.stats.percentile(
-        input,
-        50.0,
-        axis=axis,
-        interpolation="midpoint",
-        keepdims=keepdims,
-    )
-
-
-def flipud(
-    m: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    return tf.experimental.numpy.flipud(m)
-
-
-def fmod(
-    x1: Union[tf.Tensor, tf.Variable],
-    x2: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    return tf.math.floormod(x1, x2, name=None)
-
-
-@with_unsupported_dtypes(
-    {"2.9.1 and below": ("blfoat16", "float16", "float32", "float64")}, backend_version
-)
-def fmax(
-    x1: Union[tf.Tensor, tf.Variable],
-    x2: Union[tf.Tensor, tf.Variable],
-    /,
-    *,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    x1 = tf.where(tf.math.is_nan(x1), float("inf"), x1)
-    x2 = tf.where(tf.math.is_nan(x1), float("inf"), x2)
-    ret = tf.math.maximum(x1, x2, name=None)
-    return tf.where(tf.math.is_inf(ret), float("nan"))
