+++ ivy/functional/backends/tensorflow/extensions/statistical.py
@@ -1,0 +1,20 @@
+from typing import Union, Optional, Tuple
+import tensorflow as tf
+import tensorflow_probability as tfp
+
+
+def median(
+    input: Union[tf.Tensor, tf.Variable],
+    /,
+    *,
+    axis: Optional[Union[Tuple[int], int]] = None,
+    keepdims: Optional[bool] = False,
+    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
+) -> Union[tf.Tensor, tf.Variable]:
+    return tfp.stats.percentile(
+        input,
+        50.0,
+        axis=axis,
+        interpolation="midpoint",
+        keepdims=keepdims,
+    )
